refactor to include multiple pages
