+++ src/pages/DetailsPage.jsx
@@ -1,0 +1,9 @@
+function DetailsPage() {
+  return (
+    <div className="flex items-center justify-center min-h-[50vh]">
+      <h1 className="text-2xl font-semibold text-primary">A new page</h1>
+    </div>
+  );
+}
+
+export default DetailsPage;
